pyproject.toml
[tool.poetry]
name = "y-lab-menu"
version = "0.1.0"
description = ""
authors = ["eddy.di <eddy.di.fint@gmail.com>"]
readme = "README.md"
packages = [{include = "y_lab_menu"}]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = {extras = ["standard"], version = "^0.26.0"}
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
httpx = "^0.26.0"
pytest = "^7.4.4"
faker = "^22.5.1"
pre-commit = "^3.6.0"
redis = {extras = ["hiredis"], version = "^5.0.1"}
types-redis = "^4.6.0.20240106"
stubs = "^1.0.0"
loguru = "^0.7.2"
fastapi = {extras = ["all"], version = "^0.109.2"}
sqlalchemy = "^2.0.25"
[tool.pytest.ini_options]
testpaths = [
    "tests"
]
filterwarnings = [
    "ignore::UserWarning",
    # note the use of single quote below to denote "raw" strings in TOML
    'ignore::DeprecationWarning',
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
